pyproject.toml
[tool.poetry]
name = "bl60x-flash"
version = "0.1.3"
description = "Flash tool for Bouffalo Labs BL60x UART bootloader"
authors = ["Stefan Schake <stschake@gmail.com>"]
license = "MIT"

[tool.poetry.scripts]
bl60x-flash = 'bl60x_flash.main:main'

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.4"
tqdm = "^4.51.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
